--- src/components/common/form/input-password/index.ts
@@ -1,1 +1,0 @@
-export { InputPassword } from "./InputPassword";
