+++ src/components/uikit/Input/index.ts
@@ -1,1 +1,2 @@
 export { Input, type InputProps } from "./Input";
+export { InputPassword } from "./InputPassword";
